% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GO.R
\name{getGO}
\alias{getGO}
\title{Download and store the Gene Ontology}
\usage{
getGO(mode = "force")
}
\arguments{
\item{mode}{Either 'force' (default) or 'auto'. If 'force' the gene ontology
will get fetched no matter what. If 'auto' it will only get fetched if the
current local version is older than the one available on the net.}
}
\value{
This function is called for its side effects
}
\description{
This function downloads a copy of the gene ontology and formats it for quick 
access. It can optionally check the current version and only download a new
if it supersede the current one.
}
\examples{
if (interactive()) {
    getGO('auto')
}

}
\references{
\url{http://geneontology.org}
}
