% Generated by roxygen2 (5.0.0): do not edit by hand
% Please edit documentation in R/panviz.R
\docType{methods}
\name{panviz}
\alias{panviz}
\alias{panviz,character-method}
\alias{panviz,data.frame-method}
\alias{panviz,matrix-method}
\alias{panviz,pgVirtual-method}
\title{Convert pangenome into PanViz}
\usage{
panviz(object, ...)

\S4method{panviz}{character}(object, name = "name", go = "go", ec = "ec",
  ignore, location, consolidate = TRUE, showcase = FALSE, ...)

\S4method{panviz}{pgVirtual}(object, location, useDescription = TRUE,
  consolidate = TRUE, showcase = FALSE, ...)

\S4method{panviz}{matrix}(object, name, go, ec, location, consolidate = TRUE,
  showcase = FALSE, ...)

\S4method{panviz}{data.frame}(object, ...)
}
\arguments{
\item{object}{The object containing the pangenome data. If supplied as
pangenome matrix it is expected that rows are gene groups and columns are
genomes.}

\item{...}{Parameters passed along to the clustering and dimensionality
reduction functions. See details.}

\item{name}{Depending on object either the name of the column with the gene
group names or a vector of gene group names. See details.}

\item{go}{Depending on object either the name of the column with the gene
group gene ontology annotation or a vector/list of gene group ontologies.
See details.}

\item{ec}{Depending on object either the name of the column with the gene
group E.C. annotation or a vector/list of gene group enzyme numbers. See
details.}

\item{ignore}{Columns in the csv file to ignore, either given as column names
or indexes}

\item{location}{The path to write the resulting PanViz files to.}

\item{consolidate}{Logical. Should all data and code be consolidated into the
PanViz.html file or spread out to multiple files.}

\item{showcase}{Logical. Should the resulting PanViz.html be opened in the
default browser upon completion.}

\item{useDescription}{Logical. Should the description column in orgInfo
be used in favor of group names (if description is NA it falls back to group
name)}
}
\value{
NULL. This function is called for its side effects. If
\code{showcase=TRUE} the resulting PanViz.html file will be opened in the
default browser.
}
\description{
This method contains the main functionality of PanVizGenerator. It takes the
pangenome data and properly formats it, combines it with the PanViz code and
creates the PanViz files needed for the visualization. Per default everything
is consolidated into the PanViz.html file but data and code can also be
destributed into separate files. Currently pangenome data from
\code{\link[FindMyFriends:FindMyFriends-package]{FindMyFriends}} is supported
natively (And FindMyFriends support grouping based on other algorithms),
while more general use is supported by supplying a pangenome matrix along
with functional annotation of each row. A last option is to provide the file
path to a csv file containing the pangenome matrix along with the functional
annotation.
}
\details{
The calculation of mds/pca as well as the hierarchical clustering can be
controlled with the use of \code{dist} (default: 'canberra') for setting the
method used in the \code{\link[stats]{dist}} calls, \code{clust} (default:
'ward.D2') for setting the method used in the \code{\link[stats]{hclust}}
calls, \code{center} (default: TRUE) to control whether variables should be
centered prior to doing PCA and \code{scale} (default: TRUE) to control
whether scaling should be performed prior to PCA.

When using \code{panviz} with a csv file the \code{name}, \code{go} and
\code{ec} parameters should point to the columns in the csv file containing
the respective information, either by name or index. If E.C. annotation is
not given in the csv file it can be set to NA. For column with multiple
possible values (go and ec) any delimiter can be used but ',', '.', and
numbers (don't know why you would use numbers as delimiter anyway).

For panviz with a matrix or data.frame the \code{name}, \code{go} and
\code{ec} parameters should contain the actual annotation as character
vectors or lists of strings. For character vectors the same delimiting
restrictions exists as for csv files described above. \code{ec} and
\code{name} can be omitted. If \code{name} is missing the rownames of the
matrix or data.frame will be used instead - if these are not present an error
will be thrown. \code{name}, \code{go} and \code{ec} must match the number of
rows in the pangenome matrix if given.
}
\section{Methods (by class)}{
\itemize{
\item \code{character}: Method for file paths

\item \code{pgVirtual}: Method for pgVirtual subclasses from FindMyFriends

\item \code{matrix}: Method for pangenome matrix as numeric/integer matrix

\item \code{data.frame}: Method for pangenome matrix as data.frame (will be coerced
to matrix)
}}
\examples{
if(interactive()) {
    exampleFile <- system.file('extdata', 'exampleData.csv')
    panviz(exampleFile, location = tempdir(), showcase = TRUE)
}
}
\seealso{
\code{\link{PanVizGenerator}} for a shiny interface to converting
csv files.
}

