% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shinyStuff.R
\name{PanVizGenerator}
\alias{PanVizGenerator}
\title{Launch the PanVizGenerator shiny app}
\usage{
PanVizGenerator(...)
}
\arguments{
\item{...}{Parameters passed on to \code{\link[shiny]{runApp}}}
}
\value{
This function is called for its side effects.
}
\description{
This function launches a shiny based GUI that lets the user create panviz 
visualizations from csv files. The same formatting restriction as in 
\code{\link{panviz}} applies to the csv file. Furthermore the GUI also 
contains descriptions of the ideas behind the visualization as well as a 
demonstration video.
}
\examples{
if (interactive()) {
    PanVizGenerator()
}

}
\seealso{
\code{\link{panviz}} for an R API for the same functionality.
}

